feat: 코인지표 api 개발
- 코인지표값을 반환하는 api 추가
- 기존에 double로 되어있던 값들 모두 BigDecimal로 수정
- 코인지표 에러 처리 완료

+++ src/main/java/_1danhebojo/coalarm/coalarm_service/domain/dashboard/controller/response/LongShortStrengthDTO.java
@@ -5,11 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class LongShortStrengthDTO {
-    private double longRatio;
-    private double shortRatio;
+    private BigDecimal longRatio;
+    private BigDecimal shortRatio;
 }
